WIP: efcore.identity
 - add support for flex encryption/hash attributes.
 - add support for entity framework conventions.
 - add the following new tables: phone / api key / permissions / mfa policy

+++ dotnet/src/EfCore.Identity/EmailPurpose.cs
@@ -1,6 +1,6 @@
 namespace NerdyMishka.EfCore.Identity
 {
-    public enum EmailPurpose : byte
+    public enum EmailPurpose : int
     {
         Primary = 0,
         LinkedAccount = 1,
